Broadcast routing tables every 30 seconds.
This fixes issues where the initial I-Am-Router-To-Network message gets lost.
Such messages getting lost can result in significant problems when two BFR
Routers are talking to each other because BFR only sends a
"Who-Is-Router-To-Network" message *once* when it sees a NPDU with an unknown
DNET. If the initial "I-Am-Router-To-Network" message *and* the
"Who-Is-Router-To-Network" message *both* get lost (e.g. because there was no
initial connectivity between the two BFRs), then the system gets completely
stuck and never recovers.